+++ src/vip_client/classes/__init__.py
@@ -2,7 +2,7 @@
 All classes for the client.
 - VipSession: main user class. To run a VIP application on local datasets.
 - VipLauncher: to run a Vip application on datasets located on VIP servers.
-- VipCI (alpha): to run a Vip application on datasets located on CREATIS data warehouse.
+- VipGirder (alpha): to run a Vip application on datasets located on CREATIS data warehouse.
 - VipLoader (planned): to upload / download data to / from VIP servers.
 - VipLoader (planned): base class.
 """
@@ -10,6 +10,4 @@ All classes for the client.
 # Replace each class module by its class in the namespace
 from vip_client.classes.VipSession import VipSession 
 from vip_client.classes.VipLauncher import VipLauncher
-from vip_client.classes.VipCI import VipCI
-from vip_client.classes.VipLoader import VipLoader
-from vip_client.classes.VipClient import VipClient
+from vip_client.classes.VipGirder import VipGirder