+++ tests/test_VipLauncher.py
@@ -1,8 +1,6 @@
-import io
 import pytest
 from pathlib import *
 
-from vip_client.utils import vip
 from vip_client.classes import VipLauncher
 from mocked_services import mock_vip_api, mock_pathlib, mock_os